fix spacing to match PEP8 standards

+++ lightcurve.py
@@ -4,27 +4,29 @@ from matplotlib.path import Path
 from matplotlib.markers import MarkerStyle
 import itertools
 from astropy.table import Table, vstack
-import astropy.constants as const
-import astropy.units as u
 import filters
 
+
 class Arrow(Path):
-  def __init__(self, hx, hy):
-    '''Create an down-pointing arrow-shaped path.
+    def __init__(self, hx, hy):
+        """Create an down-pointing arrow-shaped path.
 
-    *hx* is the half-width of the arrow head (in units of the arrow length).
+        *hx* is the half-width of the arrow head (in units of the arrow length).
 
-    *hy* is the height of the arrow head (in units of the arrow length).'''
-    verts = [( 0 ,  0     ),
-             ( 0 , -1     ),
-             (-hx, -1 + hy),
-             ( 0 , -1     ),
-             ( hx, -1 + hy),
-             ( 0 , -1     ),
-             ( 0 ,  0     )]
-    codes = [Path.MOVETO, Path.LINETO, Path.LINETO, Path.LINETO, Path.LINETO, Path.LINETO, Path.CLOSEPOLY]
-    Path.__init__(self, verts, codes)
+        *hy* is the height of the arrow head (in units of the arrow length)."""
+        verts = [(0, 0),
+                 (0, -1),
+                 (-hx, -1 + hy),
+                 (0, -1),
+                 (hx, -1 + hy),
+                 (0, -1),
+                 (0, 0)]
+        codes = [Path.MOVETO, Path.LINETO, Path.LINETO, Path.LINETO, Path.LINETO, Path.LINETO, Path.CLOSEPOLY]
+        Path.__init__(self, verts, codes)
+
+
 arrow = Arrow(0.2, 0.3)
+
 
 class LC(Table):
     markers = {}
@@ -77,23 +79,17 @@ class LC(Table):
 
     def calcFlux(self, nondetSigmas=3):
         self.nondetSigmas = nondetSigmas
-        self['flux'], self['dflux'] = mag2flux(self['mag'], self['dmag'], self.zp(), self['nondet'], self.nondetSigmas)            
+        self['flux'], self['dflux'] = mag2flux(self['mag'], self['dmag'], self.zp(), self['nondet'], self.nondetSigmas)
 
-#def groupLC(self, *groupby):
-#    if self.lc is not None:# raise Exception('must run sn.readLC() first')
-#      if groupby:
-#        self.groupby = groupby
-#      groupby = list(set(self.groupby) & set(self.lc.keys()))
-#      if 'source' in groupby and self.lc.masked:
-#        self.lc['source'][self.lc['source'].mask] = 'other'
-#      if groupby:
-#        self.lc = self.lc.group_by(groupby)
-
-    def bin(self, delta=0.3, groupby=['filt', 'source']):
+    def bin(self, delta=0.3, groupby=None):
+        if groupby is None:
+            groupby = ['filt', 'source']
         subtabs = []
         self.groupby = list(set(groupby) & set(self.colnames))
         if self.groupby:
             grouped = self.group_by(groupby)
+        else:
+            grouped = self
         for g, k in zip(grouped.groups, grouped.groups.keys):
             mjd, flux, dflux = binflux(g['MJD'], g['flux'], g['dflux'], delta)
             binned = LC([mjd, flux, dflux], names=['MJD', 'flux', 'dflux'])
@@ -126,19 +122,20 @@ class LC(Table):
             if extinction is None:
                 extinction = self.sn.extinction
             if hostext is None:
-                extinction = self.hostext
+                hostext = self.sn.hostext
 
         self['absmag'] = self['mag'].data - dm
         for filt, A in extinction.items():
             for filtname in filters.filtdict[filt].names:
-                self['absmag'][self['filt']==filtname] -= A
+                self['absmag'][self['filt'] == filtname] -= A
         for filt, A in hostext.items():
-              for filtname in filters.filtdict[filt].names:
-                  self['absmag'][self['filt']==filtname] -= A
+            for filtname in filters.filtdict[filt].names:
+                self['absmag'][self['filt'] == filtname] -= A
 
     def calcLum(self, nondetSigmas=3):
         self.nondetSigmas = nondetSigmas
-        self['lum'], self['dlum'] = mag2flux(self['absmag'], self['dmag'], self.zp() + 90.19, self['nondet'], self.nondetSigmas)
+        self['lum'], self['dlum'] = mag2flux(self['absmag'], self['dmag'], self.zp() + 90.19, self['nondet'],
+                                             self.nondetSigmas)
 
     def findPeak(self, **criteria):
         useforpeak = ~self['nondet']
@@ -158,8 +155,6 @@ class LC(Table):
             self.sn.peakdate = np.nan
         self.sn.peakcriteria = criteria
 
-    def calcRDSP(self): calcPhase(self, rdsp=True)
-
     def calcPhase(self, rdsp=False):
         if rdsp and self.sn.peakdate is None:
             raise Exception('must run sn.findPeak() first')
@@ -171,8 +166,9 @@ class LC(Table):
             self.sn.refmjd = np.min(self.where(nondet=False)['MJD'].data)
         phase = (self['MJD'].data - self.sn.refmjd) / (1 + self.sn.z)
         self['phase'] = phase
-    
-    def plot(self, xcol='phase', ycol='absmag', offset_factor=1, color='filt', marker='source', use_lines=False, normalize=False, fillmark=True, **criteria):
+
+    def plot(self, xcol='phase', ycol='absmag', offset_factor=1, color='filt', marker='source', use_lines=False,
+             normalize=False, fillmark=True, **criteria):
         xchoices = ['phase', 'MJD']
         while xcol not in self.keys():
             xchoices.remove(xcol)
@@ -189,75 +185,82 @@ class LC(Table):
                 raise Exception('no columns found for y-axis')
         plotthese = np.tile(True, len(self))
         for key, value in criteria.items():
-          if isinstance(value, list):
-            subplotthese = np.tile(False, len(self))
-            for val in value:
-              subplotthese |= self[key] == val
-          else:
-            subplotthese = self[key] == value
-          plotthese &= subplotthese
+            if isinstance(value, list):
+                subplotthese = np.tile(False, len(self))
+                for val in value:
+                    subplotthese |= self[key] == val
+            else:
+                subplotthese = self[key] == value
+            plotthese &= subplotthese
         if np.any(plotthese):
             plottable = self[plotthese]
         else:
             return
         groupby = []
-        if color in plottable.keys(): groupby.append(color)
-        if marker in plottable.keys(): groupby.append(marker)
-        if groupby: plottable = plottable.group_by(groupby)
+        if color in plottable.keys():
+            groupby.append(color)
+        if marker in plottable.keys():
+            groupby.append(marker)
+        if groupby:
+            plottable = plottable.group_by(groupby)
         for g in plottable.groups:
-          filt = filters.filtdict[g['filt'][0]]
-          if color=='filt':
-            col = filt.color
-            mec = 'k' if filt.system == 'Johnson' else col
-          elif color=='name':
-            col = self.sn.plotcolor
-            mec = col if col not in ['w', '#FFFFFF'] else 'k'
-          elif color in plottable.keys():
-            col = colors[g[color][0]]
-            mec = col if col not in ['w', '#FFFFFF'] else 'k'
-          else:
-            col = 'k'
-            mec = 'k'
-          mfc = col if fillmark else 'none'
-          if marker=='name':
-            mark = self.sn.marker
-          elif marker in plottable.keys():
-            if g[marker][0] not in LC.markers:
-              LC.markers[g[marker][0]] = LC.othermarkers.next()
-            mark = LC.markers[g[marker][0]]
-          elif not marker:
-            mark = None
-          else:
-            mark = marker
-          if use_lines:
-            g.sort(xcol)
-          elif 'mag' in ycol:
-            yerr = g['dmag']
-          else:
-            yerr = g['d'+ycol]
-          x = g[xcol].data
-          y = g[ycol].data - filt.offset * offset_factor
-          if normalize and ycol == 'mag':
-            y -= self.sn.peakmag
-          elif normalize and ycol == 'absmag':
-            y -= self.sn.peakabsmag
-          if 'nondet' in g.keys() and marker: # don't plot if no markers used
-            plt.plot(x[g['nondet']], y[g['nondet']], marker=arrow, linestyle='none', ms=25, mec=mec)
-          if 'filt' in criteria:# and criteria['filt'] != ['R', 'r']:
-            label = self.sn.name #+ ' (' +', '.join(criteria['filt']) + ')'
-          else:
-            label = self.sn.name
-          if not use_lines:
-            plt.errorbar(x, y, yerr, color=mec, mfc=mfc, mec=mec, marker=mark, linestyle='none', label=label)
-          elif 'nondet' in g.colnames:
-            plt.plot(x[~g['nondet']], y[~g['nondet']], color=mec, mfc=mfc, mec=mec, marker=mark, label=label, linestyle=self.sn.linestyle, linewidth=self.sn.linewidth)
-            plt.plot(x[g['nondet']], y[g['nondet']], color=mec, mfc=mfc, mec=mec, marker=mark, linestyle='none')
-          else:
-            plt.plot(x, y, color=mec, mfc=mfc, mec=mec, marker=mark, label=label, linestyle=self.sn.linestyle, linewidth=self.sn.linewidth)
+            filt = filters.filtdict[g['filt'][0]]
+            if color == 'filt':
+                col = filt.color
+                mec = 'k' if filt.system == 'Johnson' else col
+            elif color == 'name':
+                col = self.sn.plotcolor
+                mec = col if col not in ['w', '#FFFFFF'] else 'k'
+            else:
+                col = 'k'
+                mec = 'k'
+            mfc = col if fillmark else 'none'
+            if marker == 'name':
+                mark = self.sn.marker
+            elif marker in plottable.keys():
+                if g[marker][0] not in LC.markers:
+                    for nextmarker in LC.othermarkers:
+                        if nextmarker not in LC.markers.values():
+                            LC.markers[g[marker][0]] = nextmarker
+                            break
+                mark = LC.markers[g[marker][0]]
+            elif not marker:
+                mark = None
+            else:
+                mark = marker
+            if use_lines:
+                g.sort(xcol)
+            elif 'mag' in ycol:
+                yerr = g['dmag']
+            else:
+                yerr = g['d' + ycol]
+            x = g[xcol].data
+            y = g[ycol].data - filt.offset * offset_factor
+            if normalize and ycol == 'mag':
+                y -= self.sn.peakmag
+            elif normalize and ycol == 'absmag':
+                y -= self.sn.peakabsmag
+            if 'nondet' in g.keys() and marker:  # don't plot if no markers used
+                plt.plot(x[g['nondet']], y[g['nondet']], marker=arrow, linestyle='none', ms=25, mec=mec)
+            if 'filt' in criteria:  # and criteria['filt'] != ['R', 'r']:
+                label = self.sn.name  # + ' (' +', '.join(criteria['filt']) + ')'
+            else:
+                label = self.sn.name
+            if not use_lines:
+                plt.errorbar(x, y, yerr, color=mec, mfc=mfc, mec=mec, marker=mark, linestyle='none', label=label)
+            elif 'nondet' in g.colnames:
+                plt.plot(x[~g['nondet']], y[~g['nondet']], color=mec, mfc=mfc, mec=mec, marker=mark, label=label,
+                         linestyle=self.sn.linestyle, linewidth=self.sn.linewidth)
+                plt.plot(x[g['nondet']], y[g['nondet']], color=mec, mfc=mfc, mec=mec, marker=mark, linestyle='none')
+            else:
+                plt.plot(x, y, color=mec, mfc=mfc, mec=mec, marker=mark, label=label, linestyle=self.sn.linestyle,
+                         linewidth=self.sn.linewidth)
 
     @classmethod
-    def read(self, filepath, format='ascii.fixed_width', fill_values=[('--', '0'), ('', '0')], **kwargs):
-        t = super(LC, self).read(filepath, format=format, fill_values=fill_values, **kwargs)
+    def read(cls, filepath, format='ascii.fixed_width', fill_values=None, **kwargs):
+        if fill_values is None:
+            fill_values = [('--', '0'), ('', '0')]
+        t = super(LC, cls).read(filepath, format=format, fill_values=fill_values, **kwargs)
         # make nondetections booleans
         if 'nondet' in t.keys():
             nondets = t['nondet'] == 'True'
@@ -266,54 +269,57 @@ class LC(Table):
         else:
             t['nondet'] = False
         # make filters strings (in case they are all the '0' filter)
-        filts = np.array(t['filt'], dtype=str)
-        t.remove_column('filt')
-        t['filt'] = filts
+        if 'filt' in t.colnames:
+            filts = np.array(t['filt'], dtype=str)
+            t.remove_column('filt')
+            t['filt'] = filts
         return t
 
+
 def flux2mag(flux, dflux=np.nan, zp=0, nondet=None, nondetSigmas=3):
-  flux = np.array(flux, copy=True)
-  dflux = np.array(dflux, copy=True)
-  if nondet is not None:
-#    nondet = np.array(nondet, copy=True)
-    flux[nondet] = nondetSigmas * dflux[nondet]
-    dflux[nondet] = np.nan
-  mag = -2.5*np.log10(flux) + zp
-  dmag = 2.5*dflux/(flux*np.log(10))
-  return mag, dmag
+    flux = np.array(flux, copy=True)
+    dflux = np.array(dflux, copy=True)
+    if nondet is not None:
+        flux[nondet] = nondetSigmas * dflux[nondet]
+        dflux[nondet] = np.nan
+    mag = -2.5 * np.log10(flux) + zp
+    dmag = 2.5 * dflux / (flux * np.log(10))
+    return mag, dmag
+
 
 def mag2flux(mag, dmag=np.nan, zp=0, nondet=None, nondetSigmas=3):
-  flux = 10**((zp - mag) / 2.5)
-  dflux = np.log(10) / 2.5 * flux * dmag
-  if nondet is not None:
-    dflux[nondet] = flux[nondet] / nondetSigmas
-    flux[nondet] = 0
-  return flux, dflux
+    flux = 10 ** ((zp - mag) / 2.5)
+    dflux = np.log(10) / 2.5 * flux * dmag
+    if nondet is not None:
+        dflux[nondet] = flux[nondet] / nondetSigmas
+        flux[nondet] = 0
+    return flux, dflux
+
 
 def binflux(time, flux, dflux, delta=0.2):
     bin_time = []
     bin_flux = []
     bin_dflux = []
     while len(flux) > 0:
-      grp = np.array(abs(time-time[0]) <= delta)
-      rest = np.logical_not(grp)
-      time_grp = time[grp]
-      flux_grp = flux[grp]
-      dflux_grp = dflux[grp]
-      if any(dflux_grp==0) or any(dflux_grp==9999) or any(np.isnan(dflux_grp)) or (np.ma.is_masked(dflux_grp) and any(dflux_grp.mask)):
-        x = np.mean(time_grp)
-        y = np.mean(flux_grp)
-        z = 0.
-      else:
-        x = np.mean(time_grp)
-        y = np.sum(flux_grp * dflux_grp**-2) / np.sum(dflux_grp**-2)
-        z = np.sum(dflux_grp**-2)**-0.5
-      bin_time.append(x) # indent these lines
-      bin_flux.append(y) # to exlude points
-      bin_dflux.append(z) # with no error
-      time = time[~grp]
-      flux = flux[~grp]
-      dflux = dflux[~grp]
+        grp = np.array(abs(time - time[0]) <= delta)
+        time_grp = time[grp]
+        flux_grp = flux[grp]
+        dflux_grp = dflux[grp]
+        if any(dflux_grp == 0) or any(dflux_grp == 9999) or any(np.isnan(dflux_grp)) or (
+                np.ma.is_masked(dflux_grp) and any(dflux_grp.mask)):
+            x = np.mean(time_grp)
+            y = np.mean(flux_grp)
+            z = 0.
+        else:
+            x = np.mean(time_grp)
+            y = np.sum(flux_grp * dflux_grp ** -2) / np.sum(dflux_grp ** -2)
+            z = np.sum(dflux_grp ** -2) ** -0.5
+        bin_time.append(x)  # indent these lines
+        bin_flux.append(y)  # to exlude points
+        bin_dflux.append(z)  # with no error
+        time = time[~grp]
+        flux = flux[~grp]
+        dflux = dflux[~grp]
     time = np.array(bin_time)
     flux = np.array(bin_flux)
     dflux = np.array(bin_dflux)
